serving/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Model serving and ML
transformers==4.35.2
torch==2.1.0
torchvision==0.16.0
bitsandbytes==0.41.3
accelerate==0.24.1
peft==0.6.2
safetensors==0.4.1
sentence-transformers==2.2.2

# Utilities and infrastructure
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
aiohttp==3.9.1
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Security and validation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Database and storage
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Medical data compliance
pydantic[email]==2.5.0
cryptography==41.0.8

# Model Version Tracking System (Phase 6)
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
mlflow>=1.20.0           # MLflow integration
wandb>=0.12.0           # Weights & Biases integration
jsonschema>=4.0.0       # JSON validation
python-dotenv>=0.19.0   # Environment configuration
black>=22.0.0           # Code formatting
flake8>=5.0.0          # Linting
mypy>=0.991            # Type checking